+++ docs/changes.md
@@ -32,6 +32,7 @@
 - improvements to input/output functionality for Assembly @ttsesm
 - added `mesh.remove_all_lines()` method
 - added keyword `Plane(edge_direction=...)` by @smoothumut
+- added `isolines(..., n=list())` option to pass specific values.
 
 
 
@@ -56,6 +57,7 @@ examples/pyplot/fit_curve2.py
 tests/issues/issue_1146.py
 tests/issues/discussion_1190.py
 tests/issues/test_sph_harm2.py
+tests/issues/issue_1218.py
 
 tests/snippets/test_interactive_plotxy.py
 tests/snippets/test_elastic_pendulum.py